+++ package.json
@@ -23,6 +23,7 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^5.15.4",
     "bootstrap": "^3.4.1",
+    "devbridge-autocomplete": "^1.4.11",
     "esbuild": "^0.21.2",
     "esbuild-plugin-less": "^1.3.4",
     "jquery": "^3.7.1",
