+++ root/static/less/style.less
@@ -85,6 +85,7 @@
 
 @import "./global.less";
 @import "./about.less";
+@import "./autocomplete.less";
 @import "./silk.less";
 @import "./diff.less";
 @import "./tablesorter.less";
@@ -109,7 +110,6 @@
 @import "./notification.less";
 @import "./footer.less";
 
-@import '../css/jquery.autocomplete.css';
 @import '../css/jquery.qtip.min.css';
 @import '../css/shCore.css';
 @import '../css/shThemeDefault.css';
